requirements.txt
tensorflow==2.13.0
numpy==1.24.3
Pillow==9.4.0
tk==0.1.0